backend/requirements.txt
# LangChain packages for LLM providers
langchain-core>=0.3.0
langchain-google-genai>=2.0.0  # For Google Gemini
langchain-groq>=0.2.0          # For Groq
langchain-cerebras>=0.1.0      # For Cerebras

# Environment variable management
python-dotenv>=1.0.0

# Optional: For accurate token counting
tiktoken>=0.5.0

# FastAPI Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
websockets>=12.0

# ==================== RAGService Dependencies ====================

# Vector Database - Qdrant
qdrant-client>=1.7.0

# Embeddings - Cohere (default provider)
cohere>=5.0.0

# Document Processing
aiofiles>=23.0.0               # Async file operations
pypdf>=4.0.0                   # PDF parsing
python-docx>=1.0.0             # DOCX parsing (optional)
# pdfplumber>=0.10.0           # Enhanced PDF table extraction (optional)

# Text Processing
langchain-text-splitters>=0.2.0  # For RecursiveCharacterTextSplitter (optional, we have custom impl)
